Added Itinal inheritance test file
The file that was added is the starting file. This file currently contains most examples of inheritance of javascript. Up to es5 not including es6 class method. It is currently in a draft stage as all the comments have not been completely gone through for completeness. As well as the code has not been yet split up and cleaned up. As well as the es6 class method will need to be added for future examples.